test(grid): Move keyboard navigation tests to a separete file #2975
