+++ projects/igniteui-angular/src/lib/grids/grid/cell.spec.ts
@@ -19,60 +19,11 @@ describe('IgxGrid - Cell component', () => {
     configureTestSuite();
 
     const CELL_CSS_CLASS = '.igx-grid__td';
-    const navigateHorizontallyToIndex = (
-        grid: IgxGridComponent,
-        cell: IgxGridCellComponent,
-        index: number) => new Promise(async (resolve) => {
-        // grid - the grid in which to navigate.
-        // cell - current cell from which the navigation will start.
-        // index - the index to which to navigate
-
-            const currIndex = cell.visibleColumnIndex;
-            const dir = currIndex < index ? 'ArrowRight' : 'ArrowLeft';
-            const nextIndex = dir === 'ArrowRight' ? currIndex + 1 : currIndex - 1;
-            const visibleColumns = grid.visibleColumns.sort((c1, c2) => c1.visibleIndex - c2.visibleIndex);
-            const nextCol = visibleColumns[nextIndex];
-            let nextCell = nextCol ? grid.getCellByColumn(0, nextCol.field) : null;
-            const keyboardEvent = new KeyboardEvent('keydown', {
-                code: dir,
-                key: dir
-            });
-            if (!cell.focused) {
-                cell.nativeElement.dispatchEvent(new Event('focus'));
-                grid.cdr.detectChanges();
-            }
-            // if index reached return
-            if (currIndex === index) { resolve(); return; }
-            // else call arrow up/down
-            // cell.nativeElement.dispatchEvent(keyboardEvent);
-            if (dir === 'ArrowRight') {
-                cell.nativeElement.dispatchEvent(keyboardEvent);
-            } else {
-                cell.nativeElement.dispatchEvent(keyboardEvent);
-            }
-
-            grid.cdr.detectChanges();
-            // if next row exists navigate next
-            if (nextCell) {
-                await wait(10);
-                grid.cdr.detectChanges();
-                navigateHorizontallyToIndex(grid, nextCell, index).then(() => { resolve(); });
-            } else {
-                // else wait for chunk to load.
-                grid.parentVirtDir.onChunkLoad.pipe(take(1)).subscribe({
-                    next: () => {
-                        grid.cdr.detectChanges();
-                        nextCell = nextCol ? grid.getCellByColumn(0, nextCol.field) : null;
-                        navigateHorizontallyToIndex(grid, nextCell, index).then(() => { resolve(); });
-                    }
-                });
-            }
-    });
+    
     beforeEach(async(() => {
         TestBed.configureTestingModule({
             declarations: [
                 DefaultGridComponent,
-                CtrlKeyKeyboardNagivationComponent,
                 VirtualGridComponent,
                 GridWithEditableColumnComponent,
                 NoColumnWidthGridComponent,
@@ -668,78 +619,6 @@ describe('IgxGrid - Cell component', () => {
         });
     });
 
-    it('keyboard navigation', (async () => {
-        const fix = TestBed.createComponent(DefaultGridComponent);
-        fix.detectChanges();
-
-        let topLeft;
-        let topRight;
-        let bottomLeft;
-        let bottomRight;
-        [topLeft, topRight, bottomLeft, bottomRight] = fix.debugElement.queryAll(By.css(CELL_CSS_CLASS));
-
-        topLeft.nativeElement.dispatchEvent(new Event('focus'));
-        await wait();
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(1);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('index');
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowdown', topLeft.nativeElement, true);
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(2);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('index');
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowright', bottomLeft.nativeElement, true);
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(2);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('value');
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowup', bottomRight.nativeElement, true);
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(1);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('value');
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowleft', topRight.nativeElement, true);
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(1);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('index');
-    }));
-
-    it('keyboard navigation - first/last cell jump with Ctrl', (async () => {
-        const fix = TestBed.createComponent(CtrlKeyKeyboardNagivationComponent);
-        fix.detectChanges();
-
-        const rv = fix.debugElement.query(By.css(CELL_CSS_CLASS));
-        const rv2 = fix.debugElement.query(By.css(`${CELL_CSS_CLASS}:last-child`));
-
-        rv.nativeElement.dispatchEvent(new Event('focus'));
-        await wait();
-        fix.detectChanges();
-
-        rv.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'Arrowright', ctrlKey: true }));
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(1);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('another');
-
-        rv2.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'Arrowleft', ctrlKey: true }));
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(1);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('index');
-    }));
-
     it('should fit last cell in the available display container when there is vertical scroll.', async(() => {
         const fix = TestBed.createComponent(VirtualGridComponent);
         fix.detectChanges();
@@ -790,41 +669,7 @@ describe('IgxGrid - Cell component', () => {
             lastCell.offsetWidth +
             verticalScroll.offsetWidth).toEqual(grid.calcWidth);
     }));
-
-    it('should scroll first row into view when pressing arrow up', (async () => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.detectChanges();
-        // the 2nd sell on the row with index 1
-        const cell = fix.debugElement.queryAll(By.css(`${CELL_CSS_CLASS}:nth-child(2)`))[1];
-
-        fix.componentInstance.scrollTop(25);
-        await wait(200);
-        fix.detectChanges();
-
-        let scrollContainer = fix.componentInstance.instance.verticalScrollContainer.dc.instance._viewContainer;
-        let scrollContainerOffset = scrollContainer.element.nativeElement.offsetTop;
-
-        expect(scrollContainerOffset).toEqual(-25);
-
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait();
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(10);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('value');
-
-        UIInteractions.triggerKeyDownEvtUponElem('arrowup', cell.nativeElement, true);
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        scrollContainer = fix.componentInstance.instance.verticalScrollContainer.dc.instance._viewContainer;
-        scrollContainerOffset = scrollContainer.element.nativeElement.offsetTop;
-
-        expect(scrollContainerOffset).toEqual(0);
-        expect(fix.componentInstance.selectedCell.value).toEqual(0);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('value');
-    }));
-
+   
     it('should not reduce the width of last pinned cell when there is vertical scroll.', fakeAsync(() => {
         const fix = TestBed.createComponent(VirtualGridComponent);
         fix.detectChanges();
@@ -851,314 +696,6 @@ describe('IgxGrid - Cell component', () => {
             expect(cell.nativeElement.clientWidth).toBeGreaterThan(100);
         });
     }));
-
-    xit('keyboard navigation - should allow navigating down in virtualized grid.', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-        const cell = grid.getCellByColumn(4, 'index');
-        cell.onFocus(new Event('focus'));
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-        // navigate down to 100th row.
-        await HelperUtils.navigateVerticallyToIndex(grid, 4, 100);
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.rowIndex).toEqual(100);
-    });
-
-   xit('keyboard navigation - should allow navigating up in virtualized grid.', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.detectChanges();
-
-        const grid = fix.componentInstance.instance;
-        grid.verticalScrollContainer.addScrollTop(5000);
-
-        await wait(200);
-        fix.detectChanges();
-
-        const cell = grid.getCellByColumn(104, 'index');
-        cell.onFocus(new Event('focus'));
-        fix.detectChanges();
-
-        await HelperUtils.navigateVerticallyToIndex(grid, 104, 0);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.rowIndex).toEqual(0);
-    });
-
-    it('keyboard navigation - should allow horizontal navigation in virtualized grid.', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        const cols = [];
-        for (let i = 0; i < 10; i++) {
-            cols.push({ field: 'col' + i });
-        }
-        fix.componentInstance.cols = cols;
-        fix.componentInstance.data = fix.componentInstance.generateData(1000);
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-        const cell = grid.getCellByColumn(0, 'col3');
-        await navigateHorizontallyToIndex(grid, cell, 9);
-        expect(fix.componentInstance.selectedCell.columnIndex).toEqual(9);
-        await navigateHorizontallyToIndex(grid, fix.componentInstance.selectedCell, 1);
-        expect(fix.componentInstance.selectedCell.columnIndex).toEqual(1);
-    });
-
-    xit('keyboard navigation - should allow horizontal navigation in virtualized grid with pinned cols.', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        const cols = [];
-        for (let i = 0; i < 10; i++) {
-            cols.push({ field: 'col' + i });
-        }
-        fix.componentInstance.cols = cols;
-        fix.componentInstance.data = fix.componentInstance.generateData(1000);
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-
-        grid.pinColumn('col1');
-        grid.pinColumn('col3');
-        fix.detectChanges();
-        const cell = grid.getCellByColumn(0, 'col1');
-        await navigateHorizontallyToIndex(grid, cell, 9);
-        expect(fix.componentInstance.selectedCell.visibleColumnIndex).toEqual(9);
-        await navigateHorizontallyToIndex(grid, fix.componentInstance.selectedCell, 1);
-        expect(fix.componentInstance.selectedCell.visibleColumnIndex).toEqual(1);
-    });
-
-    xit('keyboard navigation - should allow vertical navigation in virtualized grid with pinned cols.', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-        grid.pinColumn('index');
-
-        const cell = grid.getCellByColumn(4, 'index');
-        cell.onFocus(new Event('focus'));
-        fix.detectChanges();
-        // navigate down to 100th row.
-        await HelperUtils.navigateVerticallyToIndex(grid, 4, 100);
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.rowIndex).toEqual(100);
-    });
-
-    it('keyboard navigation - should scroll into view the not fully visible cells when navigating down', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.componentInstance.cols = fix.componentInstance.generateCols(100);
-        fix.componentInstance.data = fix.componentInstance.generateData(1000);
-        fix.detectChanges();
-
-        const grid = fix.componentInstance.instance;
-        const rows = fix.nativeElement.querySelectorAll('igx-grid-row');
-        const cell = rows[3].querySelectorAll('igx-grid-cell')[1];
-        const bottomRowHeight = rows[4].offsetHeight;
-        const displayContainer = fix.nativeElement.querySelectorAll('igx-display-container')[1];
-        const bottomCellVisibleHeight = displayContainer.parentElement.offsetHeight % bottomRowHeight;
-
-        cell.dispatchEvent(new Event('focus'));
-        await wait(50);
-
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual(30);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('1');
-        const curCell = grid.getCellByColumn(3, '1');
-        curCell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown' }));
-        await wait(50);
-
-        fix.detectChanges();
-        expect(parseInt(displayContainer.style.top, 10)).toBeLessThanOrEqual(-1 * (grid.rowHeight - bottomCellVisibleHeight));
-        expect(displayContainer.parentElement.scrollTop).toEqual(0);
-        expect(fix.componentInstance.selectedCell.value).toEqual(40);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('1');
-        });
-
-    it('keyboard navigation - should scroll into view the not fully visible cells when navigating up', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.componentInstance.cols = fix.componentInstance.generateCols(100);
-        fix.componentInstance.data = fix.componentInstance.generateData(1000);
-
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-        const displayContainer = fix.nativeElement.querySelectorAll('igx-display-container')[1];
-        fix.componentInstance.scrollTop(25);
-        await wait(50);
-
-        fix.detectChanges();
-        expect(displayContainer.style.top).toEqual('-25px');
-        const rows = fix.nativeElement.querySelectorAll('igx-grid-row');
-        const cell = rows[1].querySelectorAll('igx-grid-cell')[1];
-        cell.dispatchEvent(new Event('focus'));
-        await wait(50);
-
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual(10);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('1');
-
-        const curCell = grid.getCellByColumn(1, '1');
-        curCell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp'}));
-        await wait(50);
-        fix.detectChanges();
-
-        fix.detectChanges();
-        expect(displayContainer.style.top).toEqual('0px');
-        expect(fix.componentInstance.selectedCell.value).toEqual(0);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('1');
-    });
-
-    it('keyboard navigation - should allow navigating first/last cell in column with down/up and Cntr key.', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.detectChanges();
-
-        const grid = fix.componentInstance.instance;
-        grid.verticalScrollContainer.addScrollTop(5000);
-
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        let cell = grid.getCellByColumn(104, 'value');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(cell.selected).toBe(true);
-        expect(cell.focused).toBe(true);
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown', ctrlKey: true }));
-        await wait(200);
-        fix.detectChanges();
-
-        let selectedCellFromGrid = grid.selectedCells[0];
-        expect(fix.componentInstance.selectedCell.value).toEqual(9990);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('value');
-        expect(fix.componentInstance.selectedCell.rowIndex).toEqual(999);
-        expect(selectedCellFromGrid.value).toEqual(9990);
-        expect(selectedCellFromGrid.column.field).toMatch('value');
-        expect(selectedCellFromGrid.rowIndex).toEqual(999);
-
-        cell = grid.getCellByColumn(998, 'other');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(cell.selected).toBe(true);
-        expect(cell.focused).toBe(true);
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp', ctrlKey: true }));
-        await wait(200);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(0);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('other');
-        expect(fix.componentInstance.selectedCell.rowIndex).toEqual(0);
-        expect(grid.verticalScrollContainer.getVerticalScroll().scrollTop ).toEqual(0);
-        selectedCellFromGrid = grid.selectedCells[0];
-        expect(selectedCellFromGrid.value).toEqual(0);
-        expect(selectedCellFromGrid.column.field).toMatch('other');
-        expect(selectedCellFromGrid.rowIndex).toEqual(0);
-    });
-
-    it('keyboard navigation - should allow navigating first/last cell in column with home/end and Cntr key.', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.componentInstance.cols = fix.componentInstance.generateCols(50);
-        fix.componentInstance.data = fix.componentInstance.generateData(500);
-        fix.detectChanges();
-
-        const grid = fix.componentInstance.instance;
-        grid.verticalScrollContainer.addScrollTop(5000);
-
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        let cell = grid.getCellByColumn(101, '2');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        expect(cell.selected).toBe(true);
-        expect(cell.focused).toBe(true);
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home', ctrlKey: true }));
-        await wait(200);
-        fix.detectChanges();
-
-        let selectedCellFromGrid = grid.selectedCells[0];
-        expect(fix.componentInstance.selectedCell.value).toEqual(0);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('0');
-        expect(fix.componentInstance.selectedCell.rowIndex).toEqual(0);
-        expect(selectedCellFromGrid.value).toEqual(0);
-        expect(selectedCellFromGrid.column.field).toMatch('0');
-        expect(selectedCellFromGrid.rowIndex).toEqual(0);
-        expect(grid.verticalScrollContainer.getVerticalScroll().scrollTop ).toEqual(0);
-
-        cell = grid.getCellByColumn(4, '2');
-        cell.nativeElement.dispatchEvent(new Event('focus'));
-        await wait(DEBOUNCETIME);
-        fix.detectChanges();
-
-        cell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'End', ctrlKey: true }));
-        await wait(200);
-        fix.detectChanges();
-
-        expect(fix.componentInstance.selectedCell.value).toEqual(244510);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('49');
-        expect(fix.componentInstance.selectedCell.rowIndex).toEqual(499);
-
-        selectedCellFromGrid = grid.selectedCells[0];
-        expect(selectedCellFromGrid.value).toEqual(244510);
-        expect(selectedCellFromGrid.column.field).toMatch('49');
-        expect(selectedCellFromGrid.rowIndex).toEqual(499);
-    });
-
-    it('keyboard navigation - should scroll into view the not fully visible cells when navigating left', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.componentInstance.cols = fix.componentInstance.generateCols(100);
-        fix.componentInstance.data = fix.componentInstance.generateData(1000);
-
-        fix.detectChanges();
-        const grid = fix.componentInstance.instance;
-        const rows = fix.nativeElement.querySelectorAll('igx-grid-row');
-        const rowDisplayContainer = rows[1].querySelector('igx-display-container');
-        fix.componentInstance.scrollLeft(50);
-        await wait(50);
-        fix.detectChanges();
-
-        expect(rowDisplayContainer.style.left).toEqual('-50px');
-        const cell = rows[1].querySelectorAll('igx-grid-cell')[1];
-        cell.dispatchEvent(new Event('focus'));
-        await wait(50);
-
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual(10);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('1');
-        const curCell = grid.getCellByColumn(1, '1');
-        curCell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowLeft', ctrlKey: false }));
-        await wait(50);
-
-        fix.detectChanges();
-        expect(rowDisplayContainer.style.left).toEqual('0px');
-        expect(fix.componentInstance.selectedCell.value).toEqual(0);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('0');
-    });
-
-    it('keyboard navigation - should scroll into view the not fully visible cells when navigating right', async() => {
-        const fix = TestBed.createComponent(VirtualGridComponent);
-        fix.componentInstance.cols = fix.componentInstance.generateCols(100);
-        fix.componentInstance.data = fix.componentInstance.generateData(1000);
-        fix.detectChanges();
-
-        const grid = fix.componentInstance.instance;
-        const rows = fix.nativeElement.querySelectorAll('igx-grid-row');
-        const rowDisplayContainer = rows[1].querySelector('igx-display-container');
-        expect(rowDisplayContainer.style.left).toEqual('0px');
-        const cell = rows[1].querySelectorAll('igx-grid-cell')[2];
-        cell.dispatchEvent(new Event('focus'));
-        await wait(50);
-        fix.detectChanges();
-        expect(fix.componentInstance.selectedCell.value).toEqual(20);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('2');
-        const curCell = grid.getCellByColumn(1, '2');
-        curCell.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowRight', ctrlKey: false }));
-        await wait(50);
-
-        fix.detectChanges();
-        expect(rowDisplayContainer.style.left).toEqual('-25px');
-        expect(fix.componentInstance.selectedCell.value).toEqual(30);
-        expect(fix.componentInstance.selectedCell.column.field).toMatch('3');
-    });
 
     it('should be able to conditionally style cells', async(() => {
         const fixture = TestBed.createComponent(ConditionalCellStyleTestComponent);
@@ -1235,28 +772,6 @@ export class DefaultGridComponent {
 
     public doubleClick(evt) {
         this.clickedCell = evt.cell;
-    }
-}
-
-@Component({
-    template: `
-        <igx-grid (onSelection)="cellSelected($event)" [data]="data" [autoGenerate]="true"></igx-grid>
-    `
-})
-export class CtrlKeyKeyboardNagivationComponent {
-
-    public data = [
-        { index: 1, value: 1, other: 1, another: 1 },
-        { index: 2, value: 2, other: 2, another: 2 }
-    ];
-
-    public selectedCell: IgxGridCellComponent;
-
-    @ViewChild(IgxGridComponent, { read: IgxGridComponent })
-    public instance: IgxGridComponent;
-
-    public cellSelected(event: IGridCellEventArgs) {
-        this.selectedCell = event.cell;
     }
 }
